+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 
 ### Bug Fixes
 * Fix `SANParser` not parsing en passant captures properly (by Rob Raese).
+* Fix parsing of `"PlyCount"`, `"TimeControl"`, and `"SetUp"` PGN tags.
 
 # ChessKit 0.14.0
 Released Tuesday, June 3, 2025.